Fix budgetplan edit and get active plan
On budgetplan update I "prefill" it so that the getConflictedPlan() method can use the updated dates to find overlapping planes.
To get the active plan CURDATE() instead of NOW() should be used because NOW() also compares the time from the DateTime, which leads to non active plan on the last day. Maybe I can migrate the sql columns to be only date instead of datetime.